backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
dnspython
google-genai
transformers
torch
langgraph
langchain-google-genai
langchain-community
faiss-cpu 
sentence-transformers